loading update

+++ src/components/Products/Product/Product.js
@@ -3,30 +3,40 @@ import { useNavigate } from 'react-router-dom';
 import './Product.css'; 
 import Delete from '../../../images/logos/trash-2 9.png'; 
 import useProducts from '../../../Hooks/useProducts';
+import { useAuthState } from 'react-firebase-hooks/auth';
+import auth from '../../../firebase.init';
+import { toast } from 'react-toastify';
 
 
 const Product = ({product}) => {
   const {_id, itemName, imgLink, description, price, quantity, supplierName} = product;
   const [products, setProducts] = useProducts(); 
   const navigate = useNavigate(); 
+  const [user] = useAuthState(auth); 
   
   const navigateToProductDetails = id =>{
     navigate(`/product/${id}`);  
   }
 
   const handleProductDelete = (id) =>{
-    const proceed = window.confirm('Are you sure you want to delete this Product'); 
-    if(proceed) {
-      const url = `http://localhost:5000/product/${id}`;
-      fetch(url,{
-        method: 'DELETE'
-      })
-      .then(res=>res.json())
-      .then(data=>{  
-        const remaining = products.filter(product=>product._id !== id);
-        setProducts(remaining); 
-      })
-    } 
+    if(!user) {
+      navigate('/login'); 
+    }
+    else{
+      const proceed = window.confirm('Are you sure you want to delete this Product'); 
+      if(proceed) {
+        const url = `http://localhost:5000/product/${id}`;
+        fetch(url,{
+          method: 'DELETE'
+        })
+        .then(res=>res.json())
+        .then(data=>{  
+          toast('Successfully deleted'); 
+          const remaining = products.filter(product=>product._id !== id);
+          setProducts(remaining); 
+        })
+      } 
+    }
   }
 
   return (
